% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minqa.R
\name{print.minqa}
\alias{print.minqa}
\title{Print method for minqa objects}
\usage{
\method{print}{minqa}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object inheriting from class \code{minqa}.}

\item{digits}{unused; kept for compatibility with \code{minqa}.}

\item{...}{unused.}
}
\value{
\code{invisible(x)}
}
\description{
Print method for minqa objects
}
